+++ app/assets/javascripts/components/AccountInfo/PersonStatus/PersonStatus.js.jsx
@@ -1,0 +1,24 @@
+const React = require("react");
+
+const PersonStatus = React.createClass({
+  propTypes: {
+    person: React.PropTypes.object.isRequired,
+  },
+
+  render() {
+    const person = this.props.person;
+    return (
+      <span>
+        {(() => {
+          if (person.ready) {
+            return "(R)";
+          } else if (person.eligible) {
+            return "(E)";
+          }
+        })()}
+      </span>
+    );
+  },
+});
+
+module.exports = PersonStatus;